refactor: 선생님 부서 라벨·목록을 packages/shared로 이동
getTeacherDepartmentLabel과 TEACHER_DEPARTMENT_OPTIONS가 apps/oauth와
apps/admin 두 곳에 동일하게 정의되어 있던 중복을 제거합니다.

- 두 정의를 TeacherDepartment 타입 옆인 shared의 types/teacher.ts로 이동
- oauth/admin은 각 entities 배럴에서 shared 심볼을 재export하여 기존 import 경로 유지
- oauth의 signup/lib/utils.ts 삭제
