+++ apps/admin/src/entities/account/index.ts
@@ -1,2 +1,3 @@
 export * from './model/schema';
 export * from './lib/utils';
+export { getTeacherDepartmentLabel } from '@repo/shared/types';
