+++ apps/admin/src/entities/account/lib/utils.ts
@@ -1,9 +1,4 @@
-import {
-  AccountObjectType,
-  AccountStatus,
-  TeacherDepartment,
-  UserRoleType,
-} from '@repo/shared/types';
+import { AccountObjectType, AccountStatus, UserRoleType } from '@repo/shared/types';
 
 export const getAccountRoleLabel = (role: UserRoleType) => {
   switch (role) {
@@ -53,24 +48,5 @@ export const getAccountStatusBadgeStyle = (status: AccountStatus) => {
       return 'bg-destructive text-white border-destructive';
     case 'ACTIVE':
       return 'border-foreground/25 text-foreground';
-  }
-};
-
-export const getTeacherDepartmentLabel = (department: TeacherDepartment) => {
-  switch (department) {
-    case 'MEISTER':
-      return '마이스터부';
-    case 'DORMITORY':
-      return '사감선생님';
-    case 'GRADE':
-      return '학년부';
-    case 'ACADEMIC_AFFAIRS':
-      return '교무부';
-    case 'PROFESSIONAL_EDUCATION':
-      return '전문교육부';
-    case 'EMPLOYMENT_CAREER':
-      return '취업진로부';
-    case 'ADMINISTRATION':
-      return '행정실';
   }
 };
